backend/tournament/permissions.py
from rest_framework.permissions import BasePermission, SAFE_METHODS


class IsOrganizerOrReadOnly(BasePermission):
    def has_object_permission(self, request, view, obj):
        if request.method in SAFE_METHODS:
            return True
        tournament = getattr(obj, "tournament", obj)
        return tournament.created_by == request.user
